Update PlotCheckboxes.m make boxes smaller
checkboxes were too large for a laptop screen so you couldn't use them to turn sensors off. this update just lessens te padding between checkboxes so you can turn them on and off with a smaller screen. tested with 17" macbook screen and worked. if the screen is smaller, might need to adjust more