Fix bottom water issue
SlugHeat was depending on user defined value for "number of sensors excluding bottom water sensor" in slugpen, but most users miss this. Updated Slugheat so that whether or not a bottom water sensor is present is instead determined based on whether there is a dataset for it in the .mat file. If not, the bottom water chosen is the last measured temp in the shallowest sensor before penetration. Bottom water sensor temps are calibrated before averaged to get a single bottom water temperature.